requirements.txt.txt
httpx
python-telegram-bot==20.7
